Add support for decrypting a version-prefixed ciphertext.
Some existing implementations prefix the encrypted ciphertext with a
version byte (which is always 0x01 to my knowledge). This commit adds
support for decrypting these payloads.

This change includes an internal API to generate versioned ciphertext
for testing, but it does not add a public one. This could be done in the
future simply by exposing the existing private method.

Signed-off-by: Aaron Jacobs <aaron.jacobs@rstudio.com>