Add the "Go BoringCrypto" toolchain for CI and releases.
This commit adds a second Go toolchain to the GitHub Actions
environments used for builds, tests, and releases, as well as the
related changes to GoReleaser's configuration.

This "BoringCrypto" fork of Go's toolchain [0] is maintained by the Go
team, and has become the de-facto way to provide FIPS-validated crypto
for software written in Go. Binary releases are only available for
Linux, so that is the only platform we can support ourselves for now.

With this change, releases will have a new "rskey-fips" archive for the
linux_amd64 target. This version of rskey will refuse to use
cryptographic algorithms that are not FIPS approved, and may be broken
in unexpected ways -- some caution is warranted. However, for customers
that are in theory bound by FIPS-140, this may be a very appealing
distribution of the tool.

This commit does not include documentation for these new release
binaries because no release yet includes them.

[0]: https://go.googlesource.com/go/+/dev.boringcrypto/README.boringcrypto.md

Signed-off-by: Aaron Jacobs <aaron.jacobs@rstudio.com>