Bump to Go 1.24 and account for the new crypto/rand behaviour.
As of Go 1.24, Go itself aborts rather than return an error from
`cryto/rand`. This means that the unit tests we have for this edge case
abort immediately, and so this commit removes them.

Several internal APIs no longer need to return errors as a result,
either.

Public APIs that return an error cannot be changed, but they now
document when they will never do so.

This commit also moves us to using Go's new `tools` directive, which
should improve release automation since we no longer rely on folks to
read the `Makefile` to figure out how to install the right helper
binaries.

Signed-off-by: Aaron Jacobs <aaron.jacobs@rstudio.com>